design-patterns/src/main/java/imran/singleton/Singleton.java
package imran.singleton;

public final class Singleton {

    private static Singleton instance;
    private int value = -1;

    public static Singleton getInstance() {
        if (instance == null) {
            synchronized(Singleton.class) {
                if (instance == null) {
                    instance = new Singleton();
                }
            }
        }

        return instance;
    }

    public int getValue() {
        return value;
    }

    public void setValue(int value) {
        this.value = value;
    }
}
